[compiler] Update all dbg compile units when replacing globals
This appears to be a bug where we were assuming that `llvm.dbg.cu` only
ever contained one and only compile unit. This is probably always true
for us, but we can be more principled than this by just operating over
all operands and updating each. If there are no operands, we don't
update anything, and if there are multiple we update them all.

It also frees us up to remove a multi_llvm wrapper which was introduced
seemingly under the understanding that an empty `llvm.dbg.cu` was
invalid.

Discovered while cleaning up multi_llvm code.