[host] Relax 3.0 image error to a warning
We are in the process of removing the build-time support for different
OpenCL versions and instead moving to a unified build.

We can no longer detect 3.0 support at build time, so we can't
reasonably error if a user builds with image support.

Instead of moving the error elsewhere, we can emit a warning to users
that image support in anything but 1.2 is known to be defective.

Fixes DDK-58.