[compiler] Add !opencl.ocl.version metadata for SPIR binaries
This is required so the compiler can tell which version of OpenCL C the
module is compatible with.

We've had to determine for ourselves what that means precisely for these
binary formats.

In the case of SPIR-V we may always end up generating IR containing
sub-groups and other builtins we'd expect from OpenCL C compiled via
-cl-std=CL3.0, so we set this version. This is also regardless of SPIR-V
source language, as even for Vulkan we generate OpenCL-like IR.

In SPIR 1.2, the spec specifically makes allowances for `-cl-std`. If
the module is missing this metadata for whatever reason, we simply
assume that it's a 1.2-compatible module, as this is the default version
of `-cl-std` in the OpenCL API.

Part of DDK-112.