[UnitCL] Add shared kernel attribute for OpenCL standard
This splits out support for `-cl-std` from being specific to "clc". This
is so that kernels compiled for SPIR-V can benefit from it, without
having to abuse the system for clc-specific options, or unnecessarily
duplicating options for both output file types.

Support for SPIR has not been added (at this stage) as we are unable to
meaningfully compile kernels for OpenCL 1.2 with the OpenCL SPIRH
header and with the 3.2 version of clang we use to regenerate the SPIR
binaries.

Resolves DDK-137.