[spirv-ll] Enhance spirv-ll lit testing
This commit provides various new conveniences to the spirv-ll testing
infrastructure.

It is now structured more like our other lit test suites, with an
unconfigured lit.cfg and a configure lit.site.cfg. It works with
the `ca-lit` utility, and tests can now be run from test source paths:

    ./build/modules/lit/ca-lit modules/compiler/spirv-ll/test/spvasm/*

This automatically picks up the configured test suite location from the
build directory, mapping the source paths to the right (installed)
tests.

Running testes from source paths also means that tests can be easily
re-run without having to rebuild, if the changes don't alter the SPIR-V
binary. This is very convenient when fixing up FileCheck CHECK lines,
for instance.

In addition, we now provide a per-test substitution for the SPIR-V
binary. This simplifies the process of creating lit tests for spirv-ll
and makes it far less error prone. Previously the test would have to
provide a unique RUN line to the SPIR-V binary:

foo.spvasm:

    RUN: spirv-ll %p/foo.spvasm

This meant that copy/pasting would often result in new tests running the
incorrect SPIR-V binary. This was found in 5 of our lit tests and has
been seen previously too.

Now, using a custom "test format", each test is provided with
`%spv_file_s` which is added on a per-test basis when executing the
test. This resolves itself to the full path to the SPIR-V binary.