[host] Switch from MCJIT to ORC JIT
The `MCJIT` is a legacy code path and has been for some time. What's
more, it has some features that make it unsuitable to use safely in a
library setting: see DDK-129.

The general idea of `ORC` is basically the same as `MCJIT`; things
happen in generally the same way at the same places. The code is overall
a bit more verbose, perhaps.

One thing to note is that we must make use of "DyLibs" specific to each
instance of an optimized kernel to hold the symbols in unique
namespaces, to avoid symbol clashes.

We must also resolve all relocations in the JIT path; not just in the
offline-compiled path as before; the `MCJIT` would automatically pull in
symbols from the host process, meaning we may miss offline-only issues
like missing compiler-rt symbols. In this way, the two execution paths
are now brought closer together, which should help us in the long run.

The set-up of the JIT has been made much simpler as we no longer create
a dummy module. As a simplification of this code and convenience to
users, the `CA_HOST_TARGET_CPU` environment variable may now be set to
`native`, which does the same as when the project is built with
`-DCA_HOST_TARGET_CPU=native`.

Fixes both DDK-161 and DDK-129.