Add g1 new target template and fix external hal reference
It is simpler to create a fully working g1 target than a partial
refsi_tutorial target for demonstration purposes so added this. this
also enables the memory instrinsics so a full UnitCL can be built.

Also fixed the external HAL cmake variable name which has been hardwired
to be CPU HAL.

Additionally as part of this, the new mux target tutorial now references
the oneapi construction kit.