[muxc] Allow compiling OpenCL C to IR
This allows OpenCL C inputs to be compiled by `muxc` before running a
(now optional) pass pipeline. The input OpenCL C program is compiled
using `BaseModule::compileOpenCLCToIR`. A device is required for this to
work, as we're unable to create instances of `BaseModule` without a
known target.

The file type is worked out from the file extension, and LLVM IR is
assumed otherwise - affecting stdin, for example. The specific file type
can be set by the user, such as when piping an OpenCL C program through
stdin to the tool.

Options can be provided to this step with `-cl-options`.

This also provides the ability to select a device based on its index
from 0, using `--device-idx <idx>`. The indices are the same as those
seen when providing `--list-devices`.

There's been a fair refactoring of the code to allow this to happen, and
to tidy up the error propagation a bit.