[compiler] Move LLVMContext from BaseContext to BaseTarget
This effectively gives each target its own LLVMContext. This prepares
the way for it being a virtual method in the future, so that targets can
customize how they manage their own LLVMContexts.

It also theoretically enhances our ability to do multi-threaded
compilation across devices, though that is not specifically enabled by
this commit.