[spirv-ll] Better handle missing GLSL in lit tests
The build-tree lit.local.cfg was working when running the
`check-spirv-ll-lit` targets but not when running tests from source
paths. This is because the test suite is formally located in the source
tree, and lit.local.cfg files are only searched for in subdirectories of
the test suite.

To have this work in both cases, we must move the lit.local.cfg file and
have it feed off of build-time information, like available features.