[lit] Centralize the check for lit
This is still technically a little strong, since we don't require lit at
all when configuring or building the project - only when it comes to
running the lit tests. As such, a user could reasonably install lit
after building but before running the `check-*-lit` targets. Regardless,
it's not nice of us to generate checks (that automatically get included
in the global check target) that may fail.

With this change, at least, the duplicated code is removed, and users
can now *install* the various lit test suites, at which point users
could run them on a different machine which does have lit installed.