[lit] Don't skip suites missing tools by default
This changes the behaviour such that running lit will not by default
skip test suites if missing tools. This is much more developer friendly
as they can easily see any problems they have in test suites or in their
development environments.

When running check targets, however, we keep the old behaviour, so that
they don't cause the check to fail. This is more focused on users.