[compiler] Expose 'core' source->IR path in BaseModule
This rips out the guts of BaseModule::compileOpenCLC into a new virtual
method BaseModule::compileOpenCLCToIR, which is a 'purer' method
that compiles the source to a llvm::Module without modifying the
BaseModule's unique module. This allows testing access to this
compilation path, such as in muxc.

Note that this method isn't entirely pure as it does rely on options
having already been set on the BaseModule, and still emits diagnostics
to the BaseModule which append to its log.