[compiler] Remove LoadStoreVectorizer from our pipeline
There are a number of reasons why it doesn't make sense for us to run
this pass for targets by default:
* This is not a commonly-used pass in LLVM and only AMDGPU and NVPTX
  targets run it
* The pass doesn't handle scalable vectors, and crashes on them.
* The pass says itself that it's likely not beneficial for CPU
  architectures.
* The pass says that ARM may benefit but that ARM does its own thing.

For these reasons it doesn't make much sense to run this for all targets
by default. Targets can opt in and customize the pipeline if they wish,
or we could potentially provide a hook to do this automatically if it's
a common-enough request.