Remove requirement on spirv-tools
This is only strictly required for testing, so we can relax its
requirement somewhat.

The old fatal error message was referring to UnitCL SPIR-V tests which
isn't true; the SPIR-V tests will indeed run successfully, as they've
been pre-compiled to .spv files. It's the offline tests that won't,
because we can't detect spirv-as at build time.