[spirv-ll] Support an 'online' lit testing mode
This adds two modes to the spirv-ll lit testing infrastructure: offline
and online.

The offline mode is the default, and is essentially the current
behaviour - running `check-spirv-ll-lit` will build all the necessary
SPIR-V binaries ahead of time and run tests which expect those to be in
place.

The new online mode allows faster turnaround times on testing, and
allows the project to be built without the necessary tools known at
build time. In this mode, the tools for spvasm/glsl are checked for
independently at runtime and a feature is enabled for each if the tools
are resolved. Under this feature, a conditional RUN line is enabled to
assemble the spvasm/glsl file to SPIR-V. This uses lit's built-in `%if`
syntax.

A bonus for the online mode is that test binaries can now more easily be
assembled individually, e.g., taking certain options on a per-test
basis.

This behaviour is more intuitive for developers as iterating on
individual test files no longer requires specific build targets to be
built before the changes to the SPIR-V binary are made apparent.It also
suits building the project in a system where the tools aren't
necessarily required at build time.

Lit can be told to switch to either mode at runtime, using `--param
spirv-ll-online=1/true/0/false/`. Omitting the parameter leaves the test
suite in the mode in which it was configured at build-time.