Produce llvm cache build for use in MR and nightly testing
This adds a matrix of operating systems, llvm versions and build types
to create multiple caches which can be used in other jobs at a later
point.

This currently only does two ubuntu operating systems; windows failed
to build and this will be addressed in a later commit.

Note this must only be run on 'main' in order for MR testing to work
on a pull request.