[spirv-ll] Support the 'SubgroupSize' ExecutionMode
This commit extends our SPIR-V consumer to understand the `SubgroupSize`
execution mode. It does so to accomodate both higher-level standards
like SYCL 2020 which have a `reqd_sub_group_size` attribute and OpenCL
which has `intel_reqd_sub_group_size` (via an extension). Although
there's no mandate that this execution mode should encode this data,
that's what tools like DPC++ do in practice and so we should handle
that.

This is awkward for us, as `SubgroupSize` comes as part of the
`SubgroupDispatch` capability, which in turn implicitly defines the
`DeviceEnqueue` capability. The OpenCL environment spec explicitly says
we are not required to support `DeviceEnqueue` as we don't support
device-side enqueue, but it also says that environments supporting
SPIR-V v1.1 binaries should support `SubgroupDispatch` for OpenCL 3.0
devices supporting subgroups. This indicates a spec bug, and is
something we'll raise and hopefully fix in a future version of the
specs.

As a result, it's now perfectly possible for us to encounter SPIR-V
binaries that declare `SubgroupDispatch` and thus are implicitly allowed
to use device-side enqueue features. I feel it's better if we're strict
here and, rather than silently ignoring such features, we throw an error
to let users know what's going on. I've done that, and added a bit of
testing to cover it: full lit testing for each individual thing we don't
support would require a lot of noisy tests.

This also interops with the cl_intel_required_subgroup_size OpenCL
extension, meaning the extension should now also work with SPIR-V
programs. With that, the extension can be considered complete and safe
to enable by default. That will be done in a follow-up commit.