[spirv-ll] Fix spirv-as version feature
In offline mode, this feature wasn't including the current year for the
tools version, meaning some tests may have been unnecessarily skipped.

In online mode, this was completely broken, and was adding all years
from 0 to 2016. We need to try and execute the `spirv-as` we're going to
be using, and parse its output to find the version. This uses the same
regex we use in cmake for offline mode so should be consistent.

In both cases, rename the features to include a `+` to indicate that the
feature is set for tools older than the current ones. This mimics the
features we set for LLVM versions so should be well understood. I've
also added a `-only` feature, again like LLVM, so we can test very
explicit tools versions if need be.