Added PR testing for host and riscv m1 on ubuntu
This is required for the improvement of PR testing on github.

There are two main parts to this:
* We run a build run on main (with some testing) to seed an sccache
* We run PR testing on pull requests which will read the sccache only

Both use composite actions to avoid too much repetition for building
of the oneapi-construction-kit.

sccache uses an action which has an underlying usage of github caches.
This is tied to a branch, so seeding it has to be done on main to be
available to pull requests. Every time the build PR worflow is run,
this will introduce a new timestamped sccache, so this will have to be
managed. Currently this is only run either manually or when the workflow
files are updated.

Much of the complexity is in the composite actions which can be called
from the workflows and other composite actions:
* A common action which is uses to set the cmake and build the targets
* An action specifically for building riscv targets

Note the common action for building can mean that it overly sets cmake
variables in order to allow lots of different inputs. This may benefit
from reviewing at a future date.