[compiler] Update mangling APIs for target extension types
This updates our compiler's mangling and demangling APIs to handle
target extension types in LLVM 17+. We aren't actually able to mangle
OpenCL builtin types in LLVMs 15 or 16 as they are internall handled as
pointers to opaque structs, and with opaque pointers these are
insufficient to infer the pointee type.

This is combined with new unit testing of these APIs; though still at a
minimal level it's still more than we had before.