[compiler] Don't run the SpirFixupPass on SPIR-V
Now that spirv-ll does all of this on its own, we can be free of this
SPIR 1.2 specific pass. This makes SPIR-V input independent of most of
our fixup passes.

I've split the original ticket as there's new unresolved questions about
whether or not we need to be running the BitShiftFixupPass and
SoftwareDivisionPass on SPIR-V inputs.