[compiler] Update ImageArgumentSubstitutionPass for LLVM 17
When running this pass we require that the target extension types have
already been replaced with mux-compatible types: a pointer for images
and i32 for samplers. This means we need to do subtly different things
depending on the LLVM version.

This also schedules the target extension type replacement pass in the
defualt compiler pipeline, meaning LLVM 17 is now more-or-less ready to
go for OpenCL and SPIR-V inputs. SPIR 1.2 is known to be broken and will
be removed shortly.