[cmake] Continue to disable -Wmaybe-uninitialized
As stated in [bug report](https://gcc.gnu.org/bugzilla/show_bug.cgi?id=96564)
the warning was intended to be fixed in 11.4, then delayed to 11.5. The
issue is still open with no further input. This patch removes the upper
bound for setting `-Wno-maybe-uninitialized`, this enables building with
future GCC versions without seeing a wall of warnings of dubious value.