[compiler] Introduce mux builtins for 'group operations'
This introduces a set of builtins to represent operations like OpenCL's
`sub_group_xxx` and `work_group_xxx` and SPIR-V `OpGroupXXX` operations
in a language-agnostic way.

It also introduces 'vector group' operations which represent sub-groups
as the vectorizer currently thinks of them: as groups of work-items
being 'simulated' together via vectorized code, despite being executed
on only one invocation. These have no direct OpenCL equivalent, but can
be used by the compiler to simulate larger sub-groups than the hardware
actually provides.

Until now, the construction kit's handling of these operations has been
centered around analysis of the OpenCL builtins, which is inflexible and
brittle. We may one day want to support IR in a format not compatible
with OpenCL-like IR, for instance. It also means that the compiler is
either constrained by the OpenCL semantics of these operations.

These builtins are not yet generated in the default pipeline, but
there's a new pass which introduces them to the module by asking the
language-level `BuiltinInfo` to replace calls to its concept of group
builtins with mux's concept of these builtins. The builtins should also
be correctly wired up through the identification and get-or-declare
APIs, but are not yet able to be *defined* by mux. This will come in a
later change.