[host] Add better debugger integration with OrcJIT
This code was just copied from the
OrcV2Examples/LLJITWithGDBRegistrationListener example. It provides
a better debugging experience for jitted kernels, allowing the debugger
to step into kernels, disassemble the kernel, etc.

It also removes the need for a workaround specific to 64-bit linux
targets with LLVM 17, so the code is cleaner and more universal as a
result.