[compiler] Identify the event type for mux DMA builtins
This helps better solidify the API consistency between the
identification and declaration/definition code of overloadable builtins.

Since the mux DMA read/write builtins now require the event type to
declare themselves (as the event type is up to the target), we should be
returning that event type when being identified. The goal is to allow a
complete round-trip of identification, declaration, and definition.