[spirv-ll] Fix potential null dereference
This commit changes the behaviour in that an unknown OpConstantNull
using a target extension type will now hit an 'unreachable', whereas
before it would deference a null pointer.

This is okay since OpConstantNull is only valid on a limited range of
types (see the spec) and out of these, we'd only expect events to be
target extension types. Anything else is unhandled, as we do with
device-side events, queues, etc.