[compiler] Remove assume-like intrinsics used by allocas
In the barriers pass, when a live var is an alloca instruction, we
replace it and all of its uses directly with a GEP into the live
variable struct.

This is unsafe to do in the presence of "assume-like" intrinsics, as the
new memory region is not the same as the old one. Some interesting
examples include `llvm.assume`, whose assumption may no longer hold,
`llvm.lifetime.*` intrinsics, as the lifetime of the GEP is not equal to
that of the alloca, and `llvm.invariant.*` intrinsics.

This doesn't seem to have shown up as a problem before LLVM 17, where a
`llvm.lifetime.start` intrinsic using an alloca with the *base address*
of the live variables struct. This lifetime intrinsic was being called
after a store to *another* member of the live variables struct that we
inserted during the pass. LLVM was then later able to replace the base
GEP with the actual pointer (including in the lifetime intrinsic), and
thus infer that the store was happening before the lifetime of its base
address, and so removed the store.