[abacus] Fix abs_diff implementation
Doing the subtraction in signed types would cause the compiler to assume
no underflow, which is not a valid assumption. In C, the assumption that
there's no underflow means clang adds `nsw` flags to the subtraction(s).

It appears that LLVM 17 was finally able to make use of these flags and
optimize the IR such that our implementation was proven incorrect.