[compiler] Provide a mapping from GroupCollective -> BuiltinID
This commit allows users to get the mux builtin ID corresponding to a
`utils::GroupCollective`.

This could be useful to programmatically mutate a group collective, such
as changing a sub-group operation to a work-group one, or changing an
inclusive scan to an exclusive one.

In the absence of any concrete users, this enables round-trip testing of
`ID -> GroupCollective -> ID` in `UnitCompiler`.