[compiler] Rename HandleBarriersPass -> WorkItemLoopsPass
The historical naming has long been confusing to newcomers (so much so
that explicitly documented this fact in
doc/modules/compiler/overview.rst)

Its new name reflects the fact that it *primarily* is there to satisfy
work-group scheduling requirements by adding loops over all work-items
in the work-group. The fact that is "handles barriers" is a side-effect
of doing this correctly.