[compiler] Ensure we link against static LLVM libraries
If we are building with a clang/LLVM built with
`LLVM_LINK_LLVM_DYLIB=ON`, then the static clang libraries we want to
link `compiler-base` against pull in the dynamic "LLVM" library. This
isn't the case for any of our locally built LLVMs, but is the case in
some distributed versions of clang/LLVM, e.g., Ubuntu's.

We are unable to safely link against LLVM's dynamic library for various
reasons ultimately stemming from LLVM's reliance on global state. As an
LLVM library ourselves, we may end up with some LLVM symbols in our
libCL library with others in libLLVM. In the more concrete case of
`llvm::Any` this leads to subtle breakages, as seen in issue #109.

I can't find a nice way around this other than by explicitly linking
against the full paths to the filenames of the static libraries.

Also removes some unused LLVM CMake code in muxc.