[host] Ensure we always check llvm::Error objects
In the case that no callback was provided, we weren't checking the
`Error` objects before returning a generic error, which we must do or
else LLVM will crash.

Since we can't do anything with the message, we just use
`llvm::consumeError` which consumes it and marks it as checked without
actually doing anything.