[compiler] Fix segfaults in work-group broadcasts
The code was assuming the specific value being broadcast was always
available in the barrier/live-variables struct.

This is not the case when the value is previously analyzed as 'trivial'
and is removed for space-saving reasons.

Note that in this case it's incorrect to simply rematerialize the value
as, if it's based on the local ID, the value being broadcast will differ
for each work-item receiving the broadcast, whereas we need to retrieve
a specific (uniform) value computed by the work-item doing the
broadcasting.

Doing this 'properly' whilst maintaining the struct trimming
optimization would require either some compiler analysis to only remove
values which are not (transitively) used by work-group broadcasts, or it
would require special rematerialization logic to handle these values
which are currently unsafe to rematerialize.

Since either of these approaches would be error-prone, a quick safe fix
should be to keep the full barrier struct when any of the kernel's
barriers is a work-group broadcast.