[mux] Advertise a sub-group size of 1
This updates the list of sub-group sizes reported by both the `host` and
`riscv` targets to report `1` - the trivial sub-group size.

It appears though we have to report a non-empty list. Ignoring the
issues that degenerate sub-groups provide with this system, we should
always be able to create a sub-group size of 1, if we don't vectorize.

We won't actually always create kernels with this sub-group size: we'll
create degenerate sub-groups and we'll also vectorize, so the user can't
rely on this being the *only* sub-group size they'll see if they run a
kernel; this requires further work, and ideally some SYCL specification
clarifications.