[compiler] Add a sub-group analysis pass
This pass analyses a module's use of mux sub-group builtins. It returns
a map with info about every function *definition*, whether or not that
function uses mux sub-group builtins, as well as the set of mux builtin
IDs it uses.

It is unable to glean anything about function declarations, therefore
in such a case, this analysis should not be relied on for 100%
correctness; however, this is not generally a use-case that the oneAPI
construction kit supports.

The degenerate sub-groups pass can use this analysis to reduce some of
its own computations.

In the future the vectorizer will use this to automatically vectorize
kernels using sub-groups to a device-specific sub-group size.