Added a sycl cts profile to city runner
The standard cts / default profile of city runner has a lot of code
to handle OpenCL cts and doesn't handle SYCL CTS output very well. This
has different parsing to reflect SYCL CTS.

There are 3 main possible outputs from a test:
  test cases:  <total> |  <p> passed | <f> failed | <s> skipped | <e> failed as expected
  All tests passed : with no information on how many
  No tests ran

<total>,<p>,<f>,<s>, and <e> are all decimal digits
"test cases" should be seen as a list so it won't say 0 skipped for example

This means that there is a number of ways to look for a pass:
   If it says "All tests passed" we view it a pass
   If "No tests ran" we view it as a skipped
   If it is "test cases" then we iterate through the list:
      set pass if failed as expected
      else set fail if there are any fail items
      else set pass if a pass item
      else set skipped if skipped item.

Iterating through the list can end up with pass, fail, and skipped, but
we prioritise failed for setting status, then pass so this is okay.

Due to the way it reports (e.g. all tests passed) we don't try to
captures numbers within a single test  but just check for some form of
all passed or failed and count it as one test.

Also removed some extra parts added to the regular expression parsing
in cts.py when we are using the same profile such as "All tests passed"
and "No tests ran".