[compiler] Clarify semantics of SubgroupAnalysis pass
This clarifies and tests that the analysis does not consider the use of
mux sub-group 'setter' builtins as uses of sub-groups. This was
previously under-specified.

This behaviour is because the analysis is primarily to query the use of
sub-groups that the user can observe. The effects of the 'setter'
builtins are not apparent to the user other than through the 'getter'
builtins. They are purely to provide the scaffolding and data for these
getters.