[compiler] Preserve debug sub-programs in wrapped kernels
This changes the compiler to be more principled in leaving the
properties of wrapped kernels intact when creating new wrapper
kernels.

In this case, in `createKernelWrapperFunction` the new function was
stealing the debug sub-program from the original kernel and nullifiying
the original.

This generally worked okay, even if it was the 'wrong' thing to do, as
we usually have a 1:1 relation between wrapped and wrapper kernels.
However, if a kernel ever wanted to be wrapped by more than one wrapper,
then second wrapping process would not find any debug sub-program on the
original kernel (was it was stolen) and the debug information would be
lost. This is a serious problem as when calling an inlinable function in
a wrapper with debug info(as a wrapper generally does), the call *must*
be given a debug location with a sub-program.

This will be used in the up-coming vectorizer/sub-group changes but the
problem generally applies to the current repo, even if the problem is
not manifested.