[compiler] Clear retained nodes from artificial sub-programs
Not doing so triggers an assertion in LLVM 17. Unhelpfully, it doesn't
present a verifier error until it reaches the actual machine function
emission stage, right at the end of the compilation process.

Logically I think this makes sense - the wrapper function should not
'retain' any nodes from the wrapped function's sub-program. Those are
retained by the wrapped function itself.

This doesn't seem necessary before LLVM 17 (in that there's no
assertion), and there's no clear API with which to do it, but for
consistency's sake we now do so too.