Release any queues left when atexit is called.
This is due to an issue with dpc++ where it can leave temporary queues
unreleased. This can then have the effect on the closing down of the
device.

This is resolved by registering and deregistering the queues on creation
and deletion and then releasing any queues which still have an external
reference when `atexit` is called. This required an additional device
mutex to avoid threading issues.

This should be reviewed when https://github.com/intel/llvm/issues/11156
is fixed.