[compiler] Flag uses of half types.
When targeting a device that does not support half types, if we
nonetheless received input that used half types, we would rely on LLVM's
software implementation using library functions from libgcc/compiler-rt
which we do not provide, and receive confusing linker errors as a
result. For double types, we already flagged any use of such type as an
error. This change extends that check to also flag half types, so that
we issue an easier-to-understand error message.